init refactor of all remaining tasks
