+++ package.json
@@ -27,7 +27,7 @@
     "rush-update": "node common/scripts/install-run-rush.js update",
     "generate": "npm run rush-update",
     "bundlesize": "cd scripts && npm run bundlesize",
-    "bundlesizecollect": "node scripts/bundle-size-collect.js",
+    "bundlesizecollect": "node scripts/just.js bundle-size-collect",
     "create-component": "node scripts/create-component.js",
     "create-package": "node scripts/create-package.js",
     "checkchange": "node common/scripts/install-run-rush.js change -v",
