+++ scripts/create-package.js
@@ -1,7 +1,7 @@
 const mustache = require('mustache');
 const argv = require('yargs').argv;
 const fs = require('fs');
-const exec = require('./exec-sync');
+const exec = require('./lint-staged/exec-sync');
 const readConfig = require('./read-config');
 const writeConfig = require('./write-config');
 const path = require('path');
