+++ scripts/lint-staged/exec-sync.js
