+++ scripts/lint-staged/logging.js
