+++ scripts/tasks/api-extractor.js
@@ -1,0 +1,3 @@
+const { apiExtractorVerifyTask, apiExtractorUpdateTask } = require('just-task-preset');
+exports.verifyApiExtractor = apiExtractorVerifyTask();
+exports.updateApiExtractor = apiExtractorUpdateTask();
