--- scripts/tasks/apiExtractor.js
@@ -1,3 +1,0 @@
-const { apiExtractorVerifyTask, apiExtractorUpdateTask } = require('just-task-preset');
-exports.apiExtractor = apiExtractorVerifyTask();
-exports.updateApi = apiExtractorUpdateTask();
