+++ scripts/tasks/rig.js
@@ -1,3 +1,5 @@
+// @ts-check
+
 const { clean } = require('./clean');
 const { copy } = require('./copy');
 const { jest } = require('./jest');
@@ -6,10 +8,12 @@ const { ts } = require('./ts');
 const { tslint } = require('./tslint');
 const { webpack } = require('./webpack');
 const { outdated, selfupdate } = require('./outdated');
-const { apiExtractor, updateApi } = require('./apiExtractor');
+const { verifyApiExtractor, updateApiExtractor } = require('./api-extractor');
 const buildCodepenExamples = require('./build-codepen-examples');
 const lintImports = require('./lint-imports');
 const prettier = require('./prettier');
+const bundleSizeCollect = require('./bundle-size-collect');
+const checkForModifiedFiles = require('./check-for-modified-files');
 
 exports.rig = {
   clean,
@@ -21,9 +25,11 @@ exports.rig = {
   webpack,
   outdated,
   selfupdate,
-  apiExtractor,
-  updateApi,
+  verifyApiExtractor,
+  updateApiExtractor,
   buildCodepenExamples,
   lintImports,
-  prettier
+  prettier,
+  bundleSizeCollect,
+  checkForModifiedFiles
 };
